refactor(controller): extract scheduling and priority logic to scheduling.go
Part 5/7 of splitting the god-controller into per-concern files.
Tracks #311.

Move phase determination, Insufficient-GPU diagnostics, FIFO queue
position calculation, and priority-to-PriorityClass mapping out of
inferenceservice_controller.go into a focused scheduling.go sibling.

Moved symbols:
  const PhaseWaitingForGPU
  var priorityClassMap
  var priorityValues
  type SchedulingInfo
  (*InferenceServiceReconciler).determinePhase
  (*InferenceServiceReconciler).getPodSchedulingInfo
  (*InferenceServiceReconciler).calculateQueuePosition
  (*InferenceServiceReconciler).resolvePriorityClassName
  (*InferenceServiceReconciler).resolveEffectivePriority

Pure relocation — no signature changes, no behavior changes. Rebased
on top of #312/#313/#314/#315 which landed earlier in the split
sequence; recreated from current main rather than git-rebased due to
overlapping conflict regions in the original diff.

Verified: go vet clean, golangci-lint 0 issues, make test green,
internal/controller coverage unchanged at 83.1%.
